feat(onboard): 회원가입 시 사용자 신체 정보 받는 로직 추가

+++ sweet-choice/src/main/java/com/codeon/sweet_choice/entity/User.java
@@ -36,9 +36,9 @@ public class User {
     @Column(length = 1)
     private String state;
 
-    private Integer height;
+    private Float height;
 
-    private Integer weight;
+    private Float weight;
 
     @Column(length = 1)
     private String sex;
